Update .gitignore to include environment-specific config files
This commit updates the '.gitignore' file to include 'application-*.properites'
files in the list of ignored files. This ensures that enviroment-specific
configuration files are not committed to the repository. The 'pom.xml'
and 'application.properites' files have also been removed from the list
of ignored files, as they should be tracked by Git.