Add resource directories to pom.xml
This commit adds src/main/resources and src/test/resources directories as resource directories in the pom.xml file. The resources and testResources tags specify the locations of resource files for the main and test sources, respectively.
By adding these resource directories, the build process will include any resource files in these directories when building the application. This change is important to ensure that resource files, such as configuration files, are included in the built application.
No other changes were made in this commit.