Enable MongoDB repositories in ManagementAppApplication
This commit updates the ManagementAppApplication class to enable MongoDB repositories using the @EnableMongoRepositories annotation. The @Import annotation has been removed, as it is no longer needed. The @Value annotation has been updated to make the port field private, and a new instance of the ManagementAppApplication class is created to retrieve the port value in the setDefaultProperties method.
This change enables the use of MongoDB repositories in the application and makes the port value accessible to the setDefaultProperties method, which sets the default server port for the application. No other changes were made in this commit.