Update MongoDBConfig to use external config files
This commit updates the MongoDBConfig class to use external configuration files for the @EnableMongoRepositories and @PropertySource annotations.
The @EnableMongoRepositories annotation is added to enable Spring Data MongoDB repositories.
The @PropertySource annotation is updated to include application-external.properties in addition to application.properties.
These changes improve the management of configuration properties and make it easier to switch between different environments.
No other changes were made in this commit.