Update application.properties to allow bean definition overriding
This commit updates the application.properties file to include spring.main.allow-bean-definition-overriding=true.
This allows for bean definition overriding, which is useful when working with multiple modules that define the same bean.
In addition, the server.base-url property has been changed to http://localhost to use the local host for testing.
No other changes were made in this commit.