Updated baseUrl in index.js file
The baseUrl in the index.js file was updated from http://40.81.16.213 to http://localhost:8080.
The code for fetching user data from the backend was also updated to use the fetch API instead of jQuery AJAX calls.
Additionally, event listeners for delete, edit, and view buttons were added to the document object
to improve the performance and readability of the code.